Moved LazyLoadingGridView to a separate file
